refactor: 코드리뷰 후 수정해야할 사항 적용 (#17) - 쓸데없는 div 태그 제거 - 컴포넌트 컨벤션에 맞춰 RemainTime.js 파일 추가 - @storybook/addon-controls 삭제 - numberPad 함수 arrow function 형태로 변경(eventHandler this 바인딩 문제) - default time을 0으로 수정 - numberPad 함수 utils에 추가 - test 코드 작성 및 업데이트
