+++ package.json
@@ -30,7 +30,6 @@
     "@babel/preset-env": "^7.11.5",
     "@babel/preset-react": "^7.10.4",
     "@storybook/addon-actions": "^6.0.26",
-    "@storybook/addon-controls": "^6.0.26",
     "@storybook/addon-essentials": "^6.0.26",
     "@storybook/addon-links": "^6.0.26",
     "@storybook/react": "^6.0.26",
@@ -66,7 +65,6 @@
   },
   "dependencies": {
     "@reduxjs/toolkit": "^1.4.0",
-    "@storybook/addon-controls": "^6.0.26",
     "axios": "^0.20.0",
     "prop-types": "^15.7.2",
     "react": "^16.13.1",
